pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mtv"
version = "0.1.0"
description = "Marko TV - Simple IPTV streaming server"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
mtv = "mtv.main:main"

[tool.setuptools.packages.find]
where = ["src"]
